+++ tests/property/api_props.rs
@@ -1,4 +1,4 @@
-use dfajit::table::TransitionTable;
+use dfajit::TransitionTable;
 use dfajit::JitDfa;
 use matchkit::Match;
 use proptest::prelude::*;
@@ -20,10 +20,10 @@ proptest! {
         // Check bounds on matches
         assert!(count <= input.len());
         for i in 0..count {
-            assert!(matches[i].start() < input.len());
-            assert!(matches[i].end() <= input.len());
-            assert!(matches[i].start() < matches[i].end());
-            assert_eq!(matches[i].pattern_id(), 0);
+            assert!((matches[i].start as usize) < input.len());
+            assert!((matches[i].end as usize) <= input.len());
+            assert!(matches[i].start < matches[i].end);
+            assert_eq!(matches[i].pattern_id, 0);
         }
     }
 }
